requirements.txt
boto==2.37.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-storages==1.1.8
django-toolbelt==0.0.1
gunicorn==19.3.0
psycopg2==2.6
pyjade==3.0.0
python-google-places==1.0.0
six==1.9.0
static3==0.5.1
